Changed Memory Types to long

+++ BenchmarkCmp/DataLoaders/Concrete/JsonDataLoader.cs
@@ -27,8 +27,8 @@ namespace BenchmarkCmp.DataLoaders.Concrete
                     BranchInstructions = GetMetricValue(x, "BranchInstructions"),
                     BranchMispredictions = GetMetricValue(x, "BranchMispredictions"),
                     CacheMisses = GetMetricValue(x, "CacheMisses"),
-                    NativeAllocatedBytes = (int?)GetMetricValue(x, "AllocatedNativeMemoryDescriptor"),
-                    NativeMemoryLeak = (int?)GetMetricValue(x, "NativeMemoryLeakDescriptor")
+                    NativeAllocatedBytes = (long?)GetMetricValue(x, "AllocatedNativeMemoryDescriptor"),
+                    NativeMemoryLeak = (long?)GetMetricValue(x, "NativeMemoryLeakDescriptor")
                 })
                 .ToArray();
         }
